Rewriting the nodejs debugger in rust and minor fixes to the integration tests
Got as far as a basic implementation of the rough architecture of Padre command in place. May need some work on the command pattern or something like that yet for dispatching commands to the PadreServer or Debugger trait.

Better timeouts on integration tests, new test for ping and pings methods and fully implemented.

Added basic functionality for the connection handling.

Threaded so it handles multiple connections.