Various bug fixes and improvements.
Fixed not finding debugger from path as a check and added a line for returning data from a function in Python.

Fixed returning value in Python.

Reworking of the PADRE interface in VIM, now no longer need to restart VIM every time you want a new PADRE session.

Moved logic for app starting to VIM socket connection saving the need for PADRE to send a `started` callback.

Fixed nasty bug with stepping in other windows.

Fixing bug where it defaults to lldb when others not specified, let the `padre` command take care of this.

Adding basic funcationality to report an updated version of PADRE.