WIP: Mostly started working on LLDB and NodeJS but both are tricky at the moment.
LLDB is tricky because it's hard to know when the LLDB prompt comes up without spawning it in a PTY (and Tokio doesn't currently support this). May be able to put a process in between that does this quickly.

Node is tricky because websocket crates on Tokio 0.2 don't really exist yet